# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for a Numismatic Club.",
        "title": "Numismatic Club API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:81",
    "basePath": "/api/v1",
    "paths": {
        "/coins/": {
            "get": {
                "description": "Retrieves all coins from the collection",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Get all coins",
                "responses": {
                    "200": {
                        "description": "List of coins",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Coin"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new coin to the collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Create a new coin",
                "parameters": [
                    {
                        "description": "Coin data",
                        "name": "coin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Coin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id of the created coin",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coins/{id}/": {
            "get": {
                "description": "Retrieves a coin by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Get coin by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coin data",
                        "schema": {
                            "$ref": "#/definitions/models.Coin"
                        }
                    },
                    "400": {
                        "description": "Invalid coin ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing coin by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Update coin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated coin data",
                        "name": "coin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Coin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/utils.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid coin ID or bad request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a coin by its ID",
                "tags": [
                    "coins"
                ],
                "summary": "Delete coin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/utils.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid coin ID",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Coin": {
            "type": "object",
            "properties": {
                "condition": {
                    "description": "Состояние (например, UNC, XF, VG и т.д.)",
                    "type": "string"
                },
                "country": {
                    "description": "Страна происхождения",
                    "type": "string"
                },
                "denomination": {
                    "description": "Номинал",
                    "type": "string"
                },
                "diameter": {
                    "description": "Диаметр в миллиметрах",
                    "type": "number"
                },
                "historicalInfo": {
                    "description": "Историческая справка о монете",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "material": {
                    "description": "Материал (золото, серебро, медь, сплав)",
                    "type": "string"
                },
                "mintMark": {
                    "description": "Монетный двор (например, \"M\" для Москвы)",
                    "type": "string"
                },
                "name": {
                    "description": "Название монеты",
                    "type": "string"
                },
                "thickness": {
                    "description": "Толщина в миллиметрах",
                    "type": "number"
                },
                "value": {
                    "description": "Оценочная стоимость монеты",
                    "type": "number"
                },
                "weight": {
                    "description": "Вес в граммах",
                    "type": "number"
                },
                "year": {
                    "description": "Год выпуска",
                    "type": "integer"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "utils.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}